moved get_calalign_offsets into the utils module and fixed docstrings
