+++ astromon/utils.py
@@ -6,7 +6,9 @@ import tempfile
 from contextlib import AbstractContextManager
 from pathlib import Path
 
+import numpy as np
 import Ska.Shell
+from astropy.table import join, hstack
 
 __all__ = [
     "communicate",
@@ -15,6 +17,8 @@ __all__ = [
     "ciao_context",
     "logging_call_decorator",
     "FlowException",
+    "calalign_from_files",
+    "get_calalign_offsets",
 ]
 
 
@@ -313,6 +317,18 @@ Decorator to add logging messages at the start/end of the decorated function.
 
 
 def calalign_from_files(calalign_dir=None):
+    """
+    Get data from calalign files in `calalign_dir`.
+
+    Parameters
+    ----------
+    calalign_dir: :any:`pathlib.Path`
+        Directory where to find the calalign files.
+
+    Returns
+    -------
+    :any:`astropy.table.Table`
+    """
     import re
     from astropy.io import fits
     from cxotime import CxoTime
@@ -384,6 +400,19 @@ def calalign_from_files(calalign_dir=None):
 
 
 def get_offsets(aca_misalign):
+    """
+    Get the yag/zag offsets from the aca_misalign matrix.
+
+    Parameters
+    ----------
+    aca_misalign: np.array
+        The ACA misalignment matrices stored in calalign files as "aca_misalign"
+
+    Returns
+    -------
+    (np.array, np.array)
+        Two arrays with shape (aca_misalign.shape[1:])
+    """
     from Quaternion import Quat
     import numpy as np
 
@@ -395,3 +424,113 @@ def get_offsets(aca_misalign):
         dzs.append(-q.pitch * 3600)
         dys.append(q.yaw * 3600)
     return np.array(dys), np.array(dzs)
+
+
+def get_calalign_offsets(all_matches, ref_calalign=None, calalign_dir=None):
+    """
+    Get a table with the yag/zag offsets subtracted by the aca_misalign matrix.
+
+    The table includes two sets of offsets: the offsets used when the observation was processed,
+    and the offsets one would get when using a reference CALALIGN.
+
+    Parameters
+    ----------
+    all_matches: :any:`astropy.table.Table`
+        The table of x-ray sources. The only required columns are 'obsid' and 'x_id'
+
+    Returns
+    -------
+    :any:`astropy.table.Table`
+        A table with keys:
+
+        - calalign_dy
+        - calalign_dz
+        - aca_misalign
+        - fts_misalign
+        - calalign_version
+        - ref_calalign_dy
+        - ref_calalign_dz
+        - ref_aca_misalign
+        - ref_fts_misalign
+        - ref_calalign_version
+    """
+    # this is a table of all calalign matrices
+    calalign = calalign_from_files(calalign_dir=calalign_dir)
+    calalign.rename_columns(
+        ["dy", "dz", "caldb_version"],
+        ["calalign_dy", "calalign_dz", "calalign_version"],
+    )
+
+    # steps:
+    # - join on detector
+    # - filter out calaligns that do not correspond to the time of the match
+    # - filter out calalign versions after the desired one
+    # - group by match (obsid and x_id are unique identifiers)
+    # - take the row corresponding to the latest calibration
+    match_w_cal = join(all_matches, calalign, keys=["detector"])
+    match_w_cal = match_w_cal[
+        (match_w_cal["start"] <= match_w_cal["time"])
+        & (match_w_cal["time"] < match_w_cal["stop"])
+    ]
+    match_w_cal["cdbv"] = [
+        [int(f) for f in s.split(".")] for s in match_w_cal["caldb_version"]
+    ]
+    match_w_cal["cav"] = [
+        [int(f) for f in s.split(".")] for s in match_w_cal["calalign_version"]
+    ]
+
+    # I actually would prefer to not sort the whole table
+    # what I want is to sort the groups, which are small
+    match_w_cal.sort(keys=["obsid", "x_id", "cav", "start"], reverse=True)
+
+    if ref_calalign is None:
+        ref_calalign = max(
+            [
+                [int(f) for f in v.split(".")]
+                for v in np.unique(calalign["calalign_version"])
+            ]
+        )
+    else:
+        ref_calalign = [int(f) for f in ref_calalign.split(".")]
+
+    # the actual calalign used must be a CalDB version no later than one used to process the OBSID
+    # NOTE: cdbv's dtype is "object" because versions have different lengthd
+    # but cav's dtype is array, so need to convert to list
+    actual = match_w_cal[[r["cdbv"] >= r["cav"].tolist() for r in match_w_cal]]
+    actual = actual.group_by(["obsid", "x_id"])
+    actual = actual[actual.groups.indices[:-1]]
+
+    # the reference calalign must be a CalDB version no later than the given CalDB reference
+    reference = match_w_cal[[ref_calalign >= r["cav"].tolist() for r in match_w_cal]]
+    reference = reference.group_by(["obsid", "x_id"])
+    reference = reference[reference.groups.indices[:-1]]
+
+    cols = [
+        "calalign_dy",
+        "calalign_dz",
+        "aca_misalign",
+        "fts_misalign",
+        "calalign_version",
+    ]
+    ref_cols = ["ref_" + c for c in cols]
+    reference.rename_columns(cols, ref_cols)
+
+    # these should be true here:
+    assert len(all_matches) == len(actual)
+    assert len(all_matches) == len(reference)
+    assert np.all(reference["obsid"] == actual["obsid"])
+    assert np.all(reference["x_id"] == actual["x_id"])
+
+    result = join(
+        all_matches[["obsid", "x_id"]],
+        hstack([actual[["obsid", "x_id"] + cols], reference[ref_cols + ["since"]]]),
+        keys=["obsid", "x_id"],
+    )
+
+    assert np.all(all_matches["obsid"] == result["obsid"])
+    assert np.all(all_matches["x_id"] == result["x_id"])
+
+    # these are observations that happened after the reference calalign was added
+    result["after_caldb"] = result["since"] < all_matches["time"]
+
+    return result
